Hard is the normal law, but it is the law

+++ src/main/java/Laws/functions/DiscreteLaw.java
@@ -6,5 +6,5 @@ import Exceptions.CalculationException;
  * Created by max on 17/03/17.
  */
 public interface DiscreteLaw {
-    double getProbabiliteDeX(double x_egal_i) throws CalculationException;
+    double getProbabiliteDeX(int x_egal_i) throws CalculationException;
 }
